enhacements

+++ views/users/users.view.php
@@ -24,7 +24,7 @@ $countusers = $db->read('users');
                 <input type="text" id="search" name="search" title="Enter any part of the name" value="<?= $search ?>">
                 <select name="approved">
                     <option value="all" <?= ($approved == '') ? "SELECTED" : "" ?>>All</option>
-                    <option value="0" <?= ($approved  == '0') ? "SELECTED" : "" ?>>Unrchived</option>
+                    <option value="0" <?= ($approved  == '0') ? "SELECTED" : "" ?>>unapproved</option>
                     <option value="1" <?= ($approved  == '1') ? "SELECTED" : "" ?>>approved</option>
                 </select>
                 <select name="limit">
@@ -37,8 +37,7 @@ $countusers = $db->read('users');
                     <option value="<?= count($countusers) ?>" <?= ($limit == count($countusers)) ? "SELECTED" : "" ?>>
                         All</option>
                 </select>
-                <button type="submit" name="searchbtn" class="removebtnstyle"><img src="/views/imgs/icons/search.png"
-                        class="icon"></button>
+                <button type="submit" name="searchbtn" class="removebtnstyle"><img src="/views/imgs/icons/search.png" class="icon"></button>
             </form>
         </div>
         <!-- <form method="POST" name="multiactions" action="">
@@ -88,42 +87,35 @@ $countusers = $db->read('users');
             $users = $db->read('users', $where, $params, $limit, null, 'id');
             foreach ($users as $user) {
             ?>
-            <!-- disabe if not author or admin -->
-            <?php $disable =  authBtns($_SESSION['uid'], $user['id'], $_SESSION['role']); ?>
-            <?php $adminDisable =  isadmin($_SESSION['role']); ?>
+                <!-- disabe if not author or admin -->
+                <?php $disable =  authBtns($_SESSION['uid'], $user['id'], $_SESSION['role']); ?>
+                <?php $adminDisable =  isadmin($_SESSION['role']); ?>
 
-            <?php if ($user['approved'] == "0") {
+                <?php if ($user['approved'] == "0") {
                     $class = "linethru";
                 } else {
                     $class = "";
                 } ?>
 
-            <tr class="<?= $class ?>">
-                <td><?= $user['id']; ?></td>
-                <td>
-                    <img src="<?= ($user['image'] ?? "views/imgs/guest.png") ?>" alt="<?= $user['name']; ?>"
-                        class="userimg scale2">
-                </td>
-                <td><?= $user['user']; ?></td>
-                <td><a href="userview?id=<?= $user['id']; ?>"><?= $user['name']; ?></a></td>
-                <td><?= $user['approved'] == '1' ? 'Yes' : 'No'; ?></td>
-                <td><?= $user['joiningdate']; ?></td>
-                <td><?= $user['vacbalance']; ?></td>
+                <tr class="<?= $class ?>">
+                    <td><?= $user['id']; ?></td>
+                    <td>
+                        <img src="<?= ($user['image'] ?? "views/imgs/guest.png") ?>" alt="<?= $user['name']; ?>" class="userimg scale2">
+                    </td>
+                    <td><?= $user['user']; ?></td>
+                    <td><a href="userview?id=<?= $user['id']; ?>"><?= $user['name']; ?></a></td>
+                    <td><?= $user['approved'] == '1' ? 'Yes' : 'No'; ?></td>
+                    <td><?= $user['joiningdate']; ?></td>
+                    <td><?= $user['vacbalance']; ?></td>
 
-                <td>
-                    <a title="Edit" href="userupdate?id=<?= $user['id'] ?>" class="<?= $disable ?>"><img class="icon"
-                            src="views/imgs/icons/edit.png"></a>
-                    <a title="Delete" onclick="return confirm('Are you sure you want to Delete?')"
-                        href="userdelete?id=<?= $user['id'] ?>"
-                        class="<?= (!$adminDisable) ? "disabled" : "";  ?> "><img class="icon"
-                            src="views/imgs/icons/delete.png"></a>
-                    <?php if ($user['approved'] == 1) { ?>
-                    <a title="Block" onclick="return confirm('Are you sure you want to Block <?= $user['name'] ?> ?')"
-                        href="userblock?id=<?= $user['id'] ?>" class="<?= (!$adminDisable) ? "disabled" : "";  ?> "><img
-                            class="icon" src="views/imgs/icons/banuser.png"></a>
-                    <?php } ?>
-                </td>
-            </tr>
+                    <td>
+                        <a title="Edit" href="userupdate?id=<?= $user['id'] ?>" class="<?= $disable ?>"><img class="icon" src="views/imgs/icons/edit.png"></a>
+                        <a title="Delete" onclick="return confirm('Are you sure you want to Delete?')" href="userdelete?id=<?= $user['id'] ?>" class="<?= (!$adminDisable) ? "disabled" : "";  ?> "><img class="icon" src="views/imgs/icons/delete.png"></a>
+                        <?php if ($user['approved'] == 1) { ?>
+                            <a title="Block" onclick="return confirm('Are you sure you want to Block <?= $user['name'] ?> ?')" href="userblock?id=<?= $user['id'] ?>" class="<?= (!$adminDisable) ? "disabled" : "";  ?> "><img class="icon" src="views/imgs/icons/banuser.png"></a>
+                        <?php } ?>
+                    </td>
+                </tr>
 
             <?php } ?>
         </table>
